app/requirements.txt
Flask==3.0.3
mecs==1.0.5
pandas==2.2.2
rapidfuzz==3.9.4
numpy==2.0.0
mysql-connector-python==9.0.0
requests==2.32.3